refactoring: split base.py
