+++ ai_code_reviewer/cli.py
@@ -8,7 +8,9 @@ from typing import Tuple, Set
 import colorlog
 from langchain_community.callbacks import get_openai_callback
 
-from ai_code_reviewer.base import ProgrammingPrincipleChecker, FilePatchReview, load_principle_checkers
+from ai_code_reviewer.review import FilePatchReview
+from ai_code_reviewer.container import load_principle_checkers
+from ai_code_reviewer.reviewers.programming_principle import ProgrammingPrincipleChecker
 from ai_code_reviewer.utils import get_files_diff
